+++ resources/views/reports.blade.php
@@ -14,11 +14,11 @@ $mostrarSololectura = true;
         <tbody>
             <tr>
             <th scope="row"> <a href="listadoContratos">Listado Contratos</a></th>
-            <td>Listado para reporte mensual. Incluye los campos: Genesys ID, APELLIDO Nombre, Plan, Estado, Sistema, Comentarios: Si posee velocidad a prueba.</td>
+            <td>Listado para reporte mensual. Campos: Genesys ID, APELLIDO Nombre, Plan, Estado, Sistema, Panel, Nodo, Barrio, Comentarios: Si posee velocidad a prueba.</td>
             </tr>
             <tr>
             <th scope="row"> <a href="listadoIssues">Listado Tickets</a></th>
-            <td>Historial completo. Incluye los campos: Número de Ticket, Fecha de Apertura, Titulo, Creador, Cliente, Estado, Panel, Descripcion del Ticket.</td>
+            <td>Historial completo. Campos: Número de Ticket, Fecha de Apertura, Titulo, Creador, Cliente, Estado, Panel, Descripcion del Ticket.</td>
             </tr>
             <tr>
             <th scope="row"> <a href="listadoContratosFull">Listado Completo Full</a></th>
@@ -26,7 +26,7 @@ $mostrarSololectura = true;
             </tr>
             <tr>
             <th scope="row"> <a href="listadoContratosNoActivos">Listado Contratos de Baja</a></th>
-            <td>Listado para Contratos de baja. Incluye los campos: Genesys ID, APELLIDO Nombre, Plan, Estado, Sistema, Comentarios: Si posee velocidad a prueba.</td>
+            <td>Listado de Contratos de baja. Campos: Genesys ID, APELLIDO Nombre, Plan, Estado, Sistema, Panel, Nodo, Barrio, Comentarios: Si posee velocidad a prueba.</td>
             </tr>
         </tbody>
     </table>